@@ -1,41 +1,37 @@
 {
   "name": "@prisma/language-support",
-  "version": "19.0.57",
+  "version": "31.0.248",
   "author": "Prisma",
   "license": "Apache-2.0",
   "workspaces": [
     "packages/*"
   ],
   "devDependencies": {
-    "@actions/github": "4.0.0",
-    "@commitlint/cli": "12.0.1",
-    "@commitlint/config-conventional": "12.0.1",
-    "@commitlint/config-lerna-scopes": "12.0.1",
-    "@types/node": "14.14.35",
-    "@typescript-eslint/eslint-plugin": "4.18.0",
-    "@typescript-eslint/parser": "4.18.0",
-    "child-process": "1.0.2",
-    "eslint": "7.22.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-prettier": "3.3.1",
-    "execa": "5.0.0",
+    "@actions/github": "5.0.0",
+    "@commitlint/cli": "13.2.1",
+    "@commitlint/config-conventional": "13.2.0",
+    "@commitlint/config-lerna-scopes": "13.2.0",
+    "@types/node": "14.17.32",
+    "@typescript-eslint/eslint-plugin": "5.2.0",
+    "@typescript-eslint/parser": "5.2.0",
+    "eslint": "8.1.0",
+    "eslint-config-prettier": "8.3.0",
+    "eslint-plugin-prettier": "4.0.0",
+    "execa": "5.1.1",
     "husky": "4.3.8",
-    "jest": "26.6.3",
-    "lerna": "4.0.0",
-    "lint-staged": "10.5.4",
-    "mocha": "8.3.2",
-    "prettier": "2.2.1",
-    "semver": "7.3.4",
-    "typescript": "4.2.3",
-    "vsce": "1.87.0",
-    "vscode-test": "1.5.1",
-    "yarn": "1.22.10"
+    "jest": "27.3.1",
+    "lint-staged": "11.2.6",
+    "prettier": "2.4.1",
+    "semver": "7.3.5",
+    "typescript": "4.4.4",
+    "vsce": "1.100.2",
+    "vscode-test": "1.6.1"
   },
   "scripts": {
     "build": "tsc -b",
     "postinstall": "cd packages/language-server && npm install && cd ../vscode && npm install && cd .. && cd .. ",
-    "format": "prettier   \"packages/**/*.{ts,js,json,md}\"   --write",
-    "lint": "eslint  \"packages/**/*.{js,ts}\" --fix",
+    "format": "prettier \"packages/**/*.{ts,js,json,md}\" --write",
+    "lint": "eslint \"packages/**/*.{js,ts}\" --fix",
     "compile": "cd packages/language-server && npm run build && cd ../vscode && npm run build && cd .. && cd ..",
     "watch": "npm run build -- -w",
     "shell:lint": "shellcheck -x scripts/*.sh",
@@ -60,8 +56,5 @@
     "packages/**/*.{js,ts}": [
       "eslint --fix"
     ]
-  },
-  "dependencies": {
-    "npm-merge-driver": "2.3.6"
   }
 }
