@@ -1,8 +1,8 @@
 {
   "name": "@prisma/language-server",
-  "description": "Implementation of a language server in node.",
+  "description": "Implementation of a language server in Node.",
   "displayName": "Prisma Language Server",
-  "version": "19.0.57",
+  "version": "31.0.248",
   "author": "Prisma",
   "license": "Apache-2.0",
   "publisher": "Prisma",
@@ -10,15 +10,13 @@
     "node": ">=10"
   },
   "prisma": {
-    "version": "c1455d0b443d66b0d9db9bcb1bb9ee0d5bbc511d"
+    "enginesVersion": "25ff8286101ce9d75ed64728222d5d9beba7d46e",
+    "cliVersion": "3.4.0-dev.39"
   },
   "bin": {
     "prisma-language-server": "dist/src/bin.js"
   },
-  "repository": {
-    "type": "github",
-    "url": "git+https://github.com/prisma/vscode.git"
-  },
+  "repository": "git@github.com:prisma/language-tools.git",
   "scripts": {
     "build": "tsc",
     "watch": "npm run build -- -w",
@@ -28,29 +26,27 @@
   "main": "./dist/index.js",
   "typings": "dist/src/index",
   "dependencies": {
-    "@prisma/fetch-engine": "2.19.0-dev.47",
-    "@prisma/get-platform": "2.19.0-dev.47",
+    "@prisma/fetch-engine": "3.4.0-20.25ff8286101ce9d75ed64728222d5d9beba7d46e",
+    "@prisma/get-platform": "3.4.0-20.25ff8286101ce9d75ed64728222d5d9beba7d46e",
     "@types/js-levenshtein": "1.1.0",
-    "execa": "5.0.0",
+    "execa": "5.1.1",
     "hasha": "5.2.2",
     "js-levenshtein": "1.1.6",
-    "klona": "2.0.4",
-    "node-fetch": "2.6.1",
-    "npm": "6.14.11",
-    "p-retry": "4.4.0",
-    "pkg-dir": "5.0.0",
+    "klona": "2.0.5",
+    "node-fetch": "2.6.5",
+    "p-retry": "4.6.1",
     "rimraf": "3.0.2",
-    "tempy": "1.0.0",
+    "tempy": "1.0.1",
     "vscode-languageserver": "7.0.0",
-    "vscode-languageserver-textdocument": "1.0.1",
-    "vscode-uri": "3.0.2"
+    "vscode-languageserver-textdocument": "1.0.2"
   },
   "devDependencies": {
-    "@types/mocha": "8.2.1",
-    "@types/node-fetch": "2.5.8",
-    "@types/rimraf": "3.0.0",
-    "mocha": "8.3.2",
-    "typescript": "4.2.3"
+    "@types/mocha": "9.0.0",
+    "@types/node": "14.17.32",
+    "@types/node-fetch": "2.5.12",
+    "@types/rimraf": "3.0.2",
+    "mocha": "9.1.3",
+    "typescript": "4.4.4"
   },
   "keywords": [
     "autocomplete",
