@@ -1,76 +1,207 @@
-/**
- * Imports
- */
-import { getPlatform, Platform } from '@prisma/get-platform'
-import path from 'path'
-import fs from 'fs'
-import exec from './prisma-fmt/exec'
-const packageJson = require('../../package.json') // eslint-disable-line
+import { TextDocument, Range } from 'vscode-languageserver-textdocument'
+import { Position } from 'vscode-languageserver'
 
-/**
- * Lookup Cache
- */
-let platform: Platform | undefined
-let version: string | undefined
+export type BlockType = 'generator' | 'datasource' | 'model' | 'enum'
 
-/**
- * Lookup version
- */
-export function getVersion(): string {
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-  if (!packageJson || !packageJson.prisma || !packageJson.prisma.version) {
-    return 'latest'
+export class Block {
+  type: BlockType
+  start: Position
+  end: Position
+  name: string
+
+  constructor(type: BlockType, start: Position, end: Position, name: string) {
+    this.type = type
+    this.start = start
+    this.end = end
+    this.name = name
   }
-  // eslint-disable-next-line
-  return packageJson.prisma.version
 }
 
-/**
- * Get the exec path
- */
-export async function getBinPath(): Promise<string> {
-  platform = platform || (await getPlatform())
-  version = version || getVersion()
-  const extension = platform === 'windows' ? '.exe' : ''
-  return path.join(__dirname, `prisma-fmt.${version}${extension}`)
+export function fullDocumentRange(document: TextDocument): Range {
+  const lastLineId = document.lineCount - 1
+  return {
+    start: { line: 0, character: 0 },
+    end: { line: lastLineId, character: Number.MAX_SAFE_INTEGER },
+  }
 }
 
-/**
- * Gets the download URL for a platform
- */
-export async function getDownloadURL(): Promise<string> {
-  platform = platform || (await getPlatform())
-  version = version || getVersion()
-  const extension = platform === 'windows' ? '.exe.gz' : '.gz'
-  return `https://binaries.prisma.sh/all_commits/${version}/${platform}/prisma-fmt${extension}`
+export function getCurrentLine(document: TextDocument, line: number): string {
+  return document.getText({
+    start: { line: line, character: 0 },
+    end: { line: line, character: Number.MAX_SAFE_INTEGER },
+  })
 }
 
-export function getCLIVersion(): string {
-  // eslint-disable-next-line
-  return packageJson.dependencies['@prisma/get-platform']
+export function convertDocumentTextToTrimmedLineArray(
+  document: TextDocument,
+): Array<string> {
+  return Array(document.lineCount)
+    .fill(0)
+    .map((_, i) => getCurrentLine(document, i).trim())
 }
 
-export function binaryIsNeeded(path: string): boolean {
-  return !fs.existsSync(path)
+export function isFirstInsideBlock(
+  position: Position,
+  currentLine: string,
+): boolean {
+  if (currentLine.trim().length === 0) {
+    return true
+  }
+
+  const stringTilPosition = currentLine.slice(0, position.character)
+  const matchArray = /\w+/.exec(stringTilPosition)
+
+  if (!matchArray) {
+    return true
+  }
+  return (
+    matchArray.length === 1 &&
+    matchArray.index !== undefined &&
+    stringTilPosition.length - matchArray.index - matchArray[0].length === 0
+  )
 }
 
-export async function getBinaryVersion(path: string): Promise<string> {
-  try {
-    const output = await exec(path, ['--version'], '')
-    return output
-  } catch (errors) {
-    console.log(errors)
+export function getWordAtPosition(
+  document: TextDocument,
+  position: Position,
+): string {
+  const currentLine = getCurrentLine(document, position.line)
+
+  // search for the word's beginning and end
+  const beginning: number = currentLine
+    .slice(0, position.character + 1)
+    .search(/\S+$/)
+  const end: number = currentLine.slice(position.character).search(/\W/)
+  if (end < 0) {
     return ''
   }
+  return currentLine.slice(beginning, end + position.character)
 }
 
-export async function testBinarySuccess(path: string): Promise<boolean> {
-  // try to execute version command
-  const version = await getBinaryVersion(path)
-  if (version === '') {
-    console.log('Binary test failed. Re-attempting a download.')
-    return false
+export function getBlockAtPosition(
+  line: number,
+  lines: Array<string>,
+): Block | void {
+  let blockType = ''
+  let blockName = ''
+  let blockStart: Position = Position.create(0, 0)
+  let blockEnd: Position = Position.create(0, 0)
+  const allowedBlockIdentifiers = ['model', 'enum', 'datasource', 'generator']
+
+  // get block beginning
+  let reachedLine = false
+  for (const [key, item] of lines.reverse().entries()) {
+    const actualIndex = lines.length - 1 - key
+    if (actualIndex === line) {
+      reachedLine = true
+    }
+    if (!reachedLine) {
+      continue
+    }
+    if (
+      allowedBlockIdentifiers.some((identifier) =>
+        item.startsWith(identifier),
+      ) &&
+      item.includes('{')
+    ) {
+      const index = item.search(/\s+/)
+      blockType = ~index
+        ? (item.slice(0, index) as BlockType)
+        : (item as BlockType)
+      blockName = item.slice(blockType.length, item.length - 2).trim()
+      blockStart = Position.create(actualIndex, 0)
+      break
+    }
+    // not inside a block
+    if (item.includes('}')) {
+      lines.reverse()
+      return
+    }
   }
-  console.log('Binary test successful.')
-  return true
+  reachedLine = false
+  // get block ending
+  for (const [key, item] of lines.reverse().entries()) {
+    if (key === line) {
+      reachedLine = true
+    }
+    if (!reachedLine) {
+      continue
+    }
+    // check if block ends here
+    if (item.startsWith('}') && blockType) {
+      blockEnd = Position.create(key, 1)
+      return new Block(blockType as BlockType, blockStart, blockEnd, blockName)
+    }
+  }
+  return
+}
+
+export function getModelOrEnumBlock(
+  blockName: string,
+  lines: string[],
+): Block | void {
+  // get start position of model type
+  const results: number[] = lines
+    .map((line, index) => {
+      if (
+        (line.includes('model') && line.includes(blockName)) ||
+        (line.includes('enum') && line.includes(blockName))
+      ) {
+        return index
+      }
+    })
+    .filter((index) => index !== undefined) as number[]
+
+  if (results.length === 0) {
+    return
+  }
+
+  const foundBlocks: Block[] = results
+    .map((result) => {
+      const block = getBlockAtPosition(result, lines)
+      if (block && block.name === blockName) {
+        return block
+      }
+    })
+    .filter((block) => block !== undefined) as Block[]
+
+  if (foundBlocks.length !== 1) {
+    return
+  }
+
+  if (!foundBlocks[0]) {
+    return
+  }
+
+  return foundBlocks[0]
+}
+
+// TODO can be removed? Since it was renamed to `previewFeatures`
+export function getExperimentalFeaturesRange(
+  document: TextDocument,
+): Range | undefined {
+  const lines = convertDocumentTextToTrimmedLineArray(document)
+  const experimentalFeatures = 'experimentalFeatures'
+  let reachedStartLine = false
+  for (const [key, item] of lines.entries()) {
+    if (item.startsWith('generator') && item.includes('{')) {
+      reachedStartLine = true
+    }
+    if (!reachedStartLine) {
+      continue
+    }
+    if (reachedStartLine && item.startsWith('}')) {
+      return
+    }
+
+    if (item.startsWith(experimentalFeatures)) {
+      const startIndex = getCurrentLine(document, key).indexOf(
+        experimentalFeatures,
+      )
+      return {
+        start: { line: key, character: startIndex },
+        end: { line: key, character: startIndex + experimentalFeatures.length },
+      }
+    }
+  }
 }
