@@ -1,10 +1,7 @@
 module.exports = {
   root: true,
   parser: '@typescript-eslint/parser',
-  plugins: [
-    '@typescript-eslint', 
-    'prettier'
-  ],
+  plugins: ['@typescript-eslint', 'prettier'],
   env: {
     node: true,
     es6: true,
@@ -13,11 +10,15 @@ module.exports = {
     'plugin:@typescript-eslint/recommended',
     'plugin:@typescript-eslint/eslint-recommended',
     'plugin:prettier/recommended',
-    'plugin:@typescript-eslint/recommended-requiring-type-checking'
+    'plugin:@typescript-eslint/recommended-requiring-type-checking',
   ],
   parserOptions: {
     ecmaVersion: 2020,
     sourceType: 'module',
-    project: ['./tsconfig.json', './packages/vscode/tsconfig.json', 'packages/language-server/tsconfig.json'],
-  }
+    project: [
+      './tsconfig.json',
+      './packages/vscode/tsconfig.json',
+      'packages/language-server/tsconfig.json',
+    ],
+  },
 }
