@@ -7,7 +7,6 @@ import {
   CompletionParams,
   CompletionList,
   CompletionItem,
-  Position,
   HoverParams,
   Hover,
   CodeActionParams,
@@ -16,8 +15,19 @@ import {
   DiagnosticSeverity,
   RenameParams,
   WorkspaceEdit,
+  CompletionTriggerKind,
 } from 'vscode-languageserver'
-import { fullDocumentRange } from './provider'
+import {
+  Block,
+  convertDocumentTextToTrimmedLineArray,
+  fullDocumentRange,
+  getBlockAtPosition,
+  getCurrentLine,
+  getExperimentalFeaturesRange,
+  getModelOrEnumBlock,
+  getWordAtPosition,
+  isFirstInsideBlock,
+} from './util'
 import { TextDocument } from 'vscode-languageserver-textdocument'
 import format from './prisma-fmt/format'
 import {
@@ -26,7 +36,7 @@ import {
   getSuggestionForBlockTypes,
   getSuggestionForFirstInsideBlock,
   getSuggestionForSupportedFields,
-  getSuggestionsForInsideAttributes,
+  getSuggestionsForInsideRoundBrackets,
   positionIsAfterFieldAndType,
   isInsideAttribute,
   getSymbolBeforePosition,
@@ -51,171 +61,10 @@ import {
   isRelationField,
 } from './rename/renameUtil'
 import {
-  checkForExperimentalFeaturesUSeage,
   checkForPrisma1Model,
   greyOutIgnoredParts,
   transformLinterErrorsToDiagnostics,
 } from './diagnosticsHandler'
-
-export function getCurrentLine(document: TextDocument, line: number): string {
-  return document.getText({
-    start: { line: line, character: 0 },
-    end: { line: line, character: Number.MAX_SAFE_INTEGER },
-  })
-}
-
-export function convertDocumentTextToTrimmedLineArray(
-  document: TextDocument,
-): Array<string> {
-  return Array(document.lineCount)
-    .fill(0)
-    .map((_, i) => getCurrentLine(document, i).trim())
-}
-
-function isFirstInsideBlock(position: Position, currentLine: string): boolean {
-  if (currentLine.trim().length === 0) {
-    return true
-  }
-
-  const stringTilPosition = currentLine.slice(0, position.character)
-  const matchArray = /\w+/.exec(stringTilPosition)
-
-  if (!matchArray) {
-    return true
-  }
-  return (
-    matchArray.length === 1 &&
-    matchArray.index !== undefined &&
-    stringTilPosition.length - matchArray.index - matchArray[0].length === 0
-  )
-}
-
-export function getWordAtPosition(
-  document: TextDocument,
-  position: Position,
-): string {
-  const currentLine = getCurrentLine(document, position.line)
-
-  // search for the word's beginning and end
-  const beginning: number = currentLine
-    .slice(0, position.character + 1)
-    .search(/\S+$/)
-  const end: number = currentLine.slice(position.character).search(/\W/)
-  if (end < 0) {
-    return ''
-  }
-  return currentLine.slice(beginning, end + position.character)
-}
-
-export class Block {
-  type: string
-  start: Position
-  end: Position
-  name: string
-
-  constructor(type: string, start: Position, end: Position, name: string) {
-    this.type = type
-    this.start = start
-    this.end = end
-    this.name = name
-  }
-}
-
-export function getBlockAtPosition(
-  line: number,
-  lines: Array<string>,
-): Block | void {
-  let blockType = ''
-  let blockName = ''
-  let blockStart: Position = Position.create(0, 0)
-  let blockEnd: Position = Position.create(0, 0)
-  const allowedBlockIdentifiers = ['model', 'enum', 'datasource', 'generator']
-
-  // get block beginning
-  let reachedLine = false
-  for (const [key, item] of lines.reverse().entries()) {
-    const actualIndex = lines.length - 1 - key
-    if (actualIndex === line) {
-      reachedLine = true
-    }
-    if (!reachedLine) {
-      continue
-    }
-    if (
-      allowedBlockIdentifiers.some((identifier) =>
-        item.startsWith(identifier),
-      ) &&
-      item.includes('{')
-    ) {
-      const index = item.search(/\s+/)
-      blockType = ~index ? item.slice(0, index) : item
-      blockName = item.slice(blockType.length, item.length - 2).trim()
-      blockStart = Position.create(actualIndex, 0)
-      break
-    }
-    // not inside a block
-    if (item.includes('}')) {
-      lines.reverse()
-      return
-    }
-  }
-  reachedLine = false
-  // get block ending
-  for (const [key, item] of lines.reverse().entries()) {
-    if (key === line) {
-      reachedLine = true
-    }
-    if (!reachedLine) {
-      continue
-    }
-    // check if block ends here
-    if (item.startsWith('}')) {
-      blockEnd = Position.create(key, 1)
-      return new Block(blockType, blockStart, blockEnd, blockName)
-    }
-  }
-  return
-}
-
-export function getModelOrEnumBlock(
-  blockName: string,
-  lines: string[],
-): Block | void {
-  // get start position of model type
-  const results: number[] = lines
-    .map((line, index) => {
-      if (
-        (line.includes('model') && line.includes(blockName)) ||
-        (line.includes('enum') && line.includes(blockName))
-      ) {
-        return index
-      }
-    })
-    .filter((index) => index !== undefined) as number[]
-
-  if (results.length === 0) {
-    return
-  }
-
-  const foundBlocks: Block[] = results
-    .map((result) => {
-      const block = getBlockAtPosition(result, lines)
-      if (block && block.name === blockName) {
-        return block
-      }
-    })
-    .filter((block) => block !== undefined) as Block[]
-
-  if (foundBlocks.length !== 1) {
-    return
-  }
-
-  if (!foundBlocks[0]) {
-    return
-  }
-
-  return foundBlocks[0]
-}
 
 export async function handleDiagnosticsRequest(
   document: TextDocument,
@@ -232,7 +81,7 @@ export async function handleDiagnosticsRequest(
   if (checkForPrisma1Model(res)) {
     if (onError) {
       onError(
-        "You are currently viewing a Prisma 1 datamodel which is based on the GraphQL syntax. The current Prisma Language Server doesn't support this syntax. Please change the file extension to `.graphql` so the Prisma Language Server does not get triggered anymore.",
+        "You might currently be viewing a Prisma 1 datamodel which is based on the GraphQL syntax. The current Prisma Language Server doesn't support this syntax. If you are handling a Prisma 1 datamodel, please change the file extension to `.graphql` so the new Prisma Language Server does not get triggered anymore.",
       )
     }
   }
@@ -242,16 +91,20 @@ export async function handleDiagnosticsRequest(
     document,
   )
 
+  // TODO can be removed? Since it was renamed to `previewFeatures`
   // check for experimentalFeatures inside generator block
-  const experimentalFeaturesUsage = checkForExperimentalFeaturesUSeage(document)
-  if (experimentalFeaturesUsage) {
-    diagnostics.push({
-      severity: DiagnosticSeverity.Warning,
-      range: experimentalFeaturesUsage,
-      message:
-        "This property has been renamed to 'previewFeatures' to better communicate what they are.",
-      source: '',
-    })
+  if (document.getText().includes('experimentalFeatures')) {
+    const experimentalFeaturesRange: Range | undefined =
+      getExperimentalFeaturesRange(document)
+    if (experimentalFeaturesRange) {
+      diagnostics.push({
+        severity: DiagnosticSeverity.Warning,
+        range: experimentalFeaturesRange,
+        message:
+          "This property has been renamed to 'previewFeatures' to better communicate what they are.",
+        source: '',
+      })
+    }
   }
 
   const lines = convertDocumentTextToTrimmedLineArray(document)
@@ -337,14 +190,9 @@ export async function handleDocumentFormatting(
 ): Promise<TextEdit[]> {
   const options = params.options
 
-  return format(
-    binPath,
-    options.tabSize,
-    document.getText(),
-    onError,
-  ).then((formatted) => [
-    TextEdit.replace(fullDocumentRange(document), formatted),
-  ])
+  return format(binPath, options.tabSize, document.getText(), onError).then(
+    (formatted) => [TextEdit.replace(fullDocumentRange(document), formatted)],
+  )
 }
 
 export function handleHoverRequest(
@@ -413,6 +261,7 @@ export function handleCompletionRequest(
     !currentLineTillPosition.includes('[') &&
     symbolBeforePositionIsWhiteSpace
 
+  // datasource, generator, model or enum
   const foundBlock = getBlockAtPosition(position.line, lines)
   if (!foundBlock) {
     if (
@@ -434,8 +283,9 @@ export function handleCompletionRequest(
   }
 
   // Completion was triggered by a triggerCharacter
-  if (context?.triggerKind === 2) {
-    switch (context.triggerCharacter) {
+  // triggerCharacters defined in src/server.ts
+  if (context?.triggerKind === CompletionTriggerKind.TriggerCharacter) {
+    switch (context.triggerCharacter as '@' | '"' | '.') {
       case '@':
         if (
           !positionIsAfterFieldAndType(position, document, wordsBeforePosition)
@@ -473,14 +323,15 @@ export function handleCompletionRequest(
     case 'model':
       // check if inside attribute
       if (isInsideAttribute(currentLineUntrimmed, position, '()')) {
-        return getSuggestionsForInsideAttributes(
+        return getSuggestionsForInsideRoundBrackets(
           currentLineUntrimmed,
           lines,
+          document,
           position,
           foundBlock,
+          binPath,
         )
       }
-
       // check if type
       if (
         !positionIsAfterFieldAndType(position, document, wordsBeforePosition)
