@@ -1,6 +1,6 @@
 import { TextDocument } from 'vscode-languageserver-textdocument'
 import { handleDocumentFormatting } from '../MessageHandler'
-import { getBinPath, binaryIsNeeded, getDownloadURL } from '../util'
+import { getBinPath, binaryIsNeeded, getDownloadURL } from '../prisma-fmt/util'
 import install from '../prisma-fmt/install'
 import { TextEdit, DocumentFormattingParams } from 'vscode-languageserver'
 import * as assert from 'assert'
@@ -21,8 +21,6 @@ async function assertFormat(fixturePath: string): Promise<void> {
     document,
     binPathPrismaFmt,
   )
-
-  // TODO apply edits and compare with fixtures/correct.prisma
 
   assert.ok(formatResult.length !== 0)
 }
