@@ -1,6 +1,6 @@
 import { TextDocument } from 'vscode-languageserver-textdocument'
 import { handleDiagnosticsRequest } from '../MessageHandler'
-import { getBinPath, binaryIsNeeded, getDownloadURL } from '../util'
+import { getBinPath, binaryIsNeeded, getDownloadURL } from '../prisma-fmt/util'
 import install from '../prisma-fmt/install'
 import { Diagnostic, DiagnosticSeverity } from 'vscode-languageserver'
 import * as assert from 'assert'
@@ -20,9 +20,9 @@ async function assertLinting(
   assert.ok(diagnosticsResults.length != 0)
   expected.forEach((expectedDiagnostic, i) => {
     const actualDiagnostic = diagnosticsResults[i]
-    assert.equal(actualDiagnostic.message, expectedDiagnostic.message)
-    assert.deepEqual(actualDiagnostic.range, expectedDiagnostic.range)
-    assert.equal(actualDiagnostic.severity, expectedDiagnostic.severity)
+    assert.strictEqual(actualDiagnostic.message, expectedDiagnostic.message)
+    assert.deepStrictEqual(actualDiagnostic.range, expectedDiagnostic.range)
+    assert.strictEqual(actualDiagnostic.severity, expectedDiagnostic.severity)
   })
 }
 
@@ -42,7 +42,6 @@ suite('Linting', () => {
 
   const fixturePathMissingArgument = './linting/missingArgument.prisma'
   const fixturePathWrongType = './linting/wrongType.prisma'
-  const fixturePathRequiredField = './linting/requiredField.prisma'
 
   test('Missing argument', async () => {
     await assertLinting(
@@ -67,28 +66,12 @@ suite('Linting', () => {
             'Type "Use" is neither a built-in type, nor refers to another model, custom type, or enum.',
           range: {
             start: { line: 14, character: 12 },
-            end: { line: 14, character: 16 },
+            end: { line: 14, character: 15 },
           },
           severity: DiagnosticSeverity.Error,
         },
       ],
       fixturePathWrongType,
-    )
-  })
-  test('Required field', async () => {
-    await assertLinting(
-      [
-        {
-          message:
-            'Error validating: The relation field `author` uses the scalar fields authorId. At least one of those fields is required. Hence the relation field must be required as well.',
-          range: {
-            start: { line: 14, character: 2 },
-            end: { line: 15, character: 0 },
-          },
-          severity: DiagnosticSeverity.Error,
-        },
-      ],
-      fixturePathRequiredField,
     )
   })
 })
