@@ -6,6 +6,7 @@ import {
   CodeActionContext,
   Command,
   commands,
+  ExtensionContext,
   Range,
   TextDocument,
   window,
@@ -15,12 +16,20 @@ import {
   CodeAction as lsCodeAction,
   CodeActionParams,
   CodeActionRequest,
-  LanguageClient,
   LanguageClientOptions,
   ProvideCodeActionsSignature,
+} from 'vscode-languageclient'
+import {
+  LanguageClient,
   ServerOptions,
   TransportKind,
 } from 'vscode-languageclient/node'
+import {
+  BinaryStorage,
+  checkAndAskForBinaryExecution,
+  printBinaryCheckWarning,
+  PRISMA_ALLOWED_BINS_KEY,
+} from '../../binaryValidator'
 import TelemetryReporter from '../../telemetryReporter'
 import {
   applySnippetWorkspaceEdit,
@@ -28,6 +37,8 @@ import {
   checkForOtherPrismaExtension,
   isDebugOrTestSession,
   isSnippetEdit,
+  restartClient,
+  createLanguageServer,
 } from '../../util'
 import { PrismaVSCodePlugin } from '../types'
 
@@ -41,16 +52,18 @@ let watcher: chokidar.FSWatcher
 const isDebugMode = () => process.env.VSCODE_DEBUG_MODE === 'true'
 const isE2ETestOnPullRequest = () => process.env.localLSP === 'true'
 
-function createLanguageServer(
+const activateClient = (
+  context: ExtensionContext,
   serverOptions: ServerOptions,
   clientOptions: LanguageClientOptions,
-): LanguageClient {
-  return new LanguageClient(
-    'prisma',
-    'Prisma Language Server',
-    serverOptions,
-    clientOptions,
-  )
+) => {
+  // Create the language client
+  client = createLanguageServer(serverOptions, clientOptions)
+
+  const disposable = client.start()
+
+  // Start the client. This will also launch the server
+  context.subscriptions.push(disposable)
 }
 
 const plugin: PrismaVSCodePlugin = {
@@ -58,17 +71,28 @@ const plugin: PrismaVSCodePlugin = {
   enabled: () => true,
   activate: async (context) => {
     const isDebugOrTest = isDebugOrTestSession()
-    const rootPath = workspace.rootPath
+    const rootPath = workspace.workspaceFolders?.[0].uri.path
 
     if (rootPath) {
       watcher = chokidar.watch(
         path.join(rootPath, '**/node_modules/.prisma/client/index.d.ts'),
         {
-          usePolling: false,
+          // ignore dotfiles (except .prisma) adjusted from chokidar README example
+          ignored: /(^|[\/\\])\.(?!prisma)./,
+          // limits how many levels of subdirectories will be traversed.
+          // Note that `node_modules/.prisma/client/` counts for 3 already
+          // Example
+          // If vs code extension is open in root folder of a project and the path to index.d.ts is
+          // ./server/database/node_modules/.prisma/client/index.d.ts
+          // then the depth is equal to 2 + 3 = 5
+          depth: 9,
+          // When false, only the symlinks themselves will be watched for changes
+          // instead of following the link references and bubbling events through the link's path.
           followSymlinks: false,
         },
       )
     }
+
     if (isDebugMode() || isE2ETestOnPullRequest()) {
       // use LSP from folder for debugging
       console.log('Using local LSP')
@@ -76,7 +100,7 @@ const plugin: PrismaVSCodePlugin = {
         path.join('../../packages/language-server/dist/src/bin'),
       )
     } else {
-      console.log('Using published LSP.')
+      console.log('Using published LSP')
       // use published npm package for production
       serverModule = require.resolve('@prisma/language-server/dist/src/bin')
     }
@@ -113,21 +137,20 @@ const plugin: PrismaVSCodePlugin = {
           _: ProvideCodeActionsSignature,
         ) {
           const params: CodeActionParams = {
-            textDocument: client.code2ProtocolConverter.asTextDocumentIdentifier(
-              document,
-            ),
+            textDocument:
+              client.code2ProtocolConverter.asTextDocumentIdentifier(document),
             range: client.code2ProtocolConverter.asRange(range),
             context: client.code2ProtocolConverter.asCodeActionContext(context),
           }
+
           return client.sendRequest(CodeActionRequest.type, params, token).then(
             (values) => {
               if (values === null) return undefined
               const result: (CodeAction | Command)[] = []
               for (const item of values) {
                 if (lsCodeAction.is(item)) {
-                  const action = client.protocol2CodeConverter.asCodeAction(
-                    item,
-                  )
+                  const action =
+                    client.protocol2CodeConverter.asCodeAction(item)
                   if (
                     isSnippetEdit(
                       item,
@@ -157,21 +180,70 @@ const plugin: PrismaVSCodePlugin = {
         },
       } as any,
     }
+    const config = workspace.getConfiguration('prisma')
+    const allowedBins = context.globalState.get<BinaryStorage>(
+      PRISMA_ALLOWED_BINS_KEY,
+    )
 
-    // Create the language client
-    client = createLanguageServer(serverOptions, clientOptions)
+    workspace.onDidChangeConfiguration(
+      async (e) => {
+        const binChanged = e.affectsConfiguration('prisma.prismaFmtBinPath')
+        if (binChanged) {
+          client = await restartClient(
+            context,
+            client,
+            serverOptions,
+            clientOptions,
+          )
+        }
+      },
+      null,
+      context.subscriptions,
+    )
 
-    const disposable = client.start()
-
-    // Start the client. This will also launch the server
-    context.subscriptions.push(disposable)
-
+    const launchAllowed = await checkAndAskForBinaryExecution(
+      context,
+      config.get('prismaFmtBinPath'),
+      allowedBins,
+    )
     context.subscriptions.push(
+      commands.registerCommand('prisma.resetAllBinDecisions', async () => {
+        await context.globalState.update(PRISMA_ALLOWED_BINS_KEY, { libs: {} })
+        client = await restartClient(
+          context,
+          client,
+          serverOptions,
+          clientOptions,
+        )
+      }),
+      commands.registerCommand(
+        'prisma.resetCurrentFmtBinDecision',
+        async () => {
+          const decisions = context.globalState.get<BinaryStorage>(
+            PRISMA_ALLOWED_BINS_KEY,
+          )
+          const currentBin: string | undefined = workspace
+            .getConfiguration('prisma')
+            .get('prismaFmtBinPath')
+          if (decisions && currentBin && currentBin !== '') {
+            delete decisions.libs[currentBin]
+          }
+          await context.globalState.update(PRISMA_ALLOWED_BINS_KEY, decisions)
+          client = await restartClient(
+            context,
+            client,
+            serverOptions,
+            clientOptions,
+          )
+        },
+      ),
       commands.registerCommand('prisma.restartLanguageServer', async () => {
-        await client.stop()
-        client = createLanguageServer(serverOptions, clientOptions)
-        context.subscriptions.push(client.start())
-        await client.onReady()
+        client = await restartClient(
+          context,
+          client,
+          serverOptions,
+          clientOptions,
+        )
         window.showInformationMessage('Prisma language server restarted.') // eslint-disable-line @typescript-eslint/no-floating-promises
       }),
       commands.registerCommand(
@@ -180,20 +252,31 @@ const plugin: PrismaVSCodePlugin = {
       ),
     )
 
+    if (launchAllowed) {
+      activateClient(context, serverOptions, clientOptions)
+    } else {
+      await printBinaryCheckWarning()
+    }
+
     if (!isDebugOrTest) {
       // eslint-disable-next-line
       const extensionId = 'prisma.' + packageJson.name
       // eslint-disable-next-line
       const extensionVersion = packageJson.version
+
       telemetry = new TelemetryReporter(extensionId, extensionVersion)
+
       context.subscriptions.push(telemetry)
+
       await telemetry.sendTelemetryEvent()
+
       if (extensionId === 'prisma.prisma-insider') {
         checkForOtherPrismaExtension()
       }
     }
 
     checkForMinimalColorTheme()
+
     if (watcher) {
       watcher.on('change', (path) => {
         console.log(`File ${path} has been changed. Restarting TS Server.`)
@@ -205,9 +288,11 @@ const plugin: PrismaVSCodePlugin = {
     if (!client) {
       return undefined
     }
+
     if (!isDebugOrTestSession()) {
       telemetry.dispose() // eslint-disable-line @typescript-eslint/no-floating-promises
     }
+
     return client.stop()
   },
 }
