@@ -10,20 +10,19 @@ async function testDiagnostics(
 
   const actualDiagnostics = vscode.languages.getDiagnostics(docUri)
 
-  assert.equal(actualDiagnostics.length, expectedDiagnostics.length)
+  assert.strictEqual(actualDiagnostics.length, expectedDiagnostics.length)
 
   expectedDiagnostics.forEach((expectedDiagnostic, i) => {
     const actualDiagnostic = actualDiagnostics[i]
-    assert.equal(actualDiagnostic.message, expectedDiagnostic.message)
-    assert.deepEqual(actualDiagnostic.range, expectedDiagnostic.range)
-    assert.equal(actualDiagnostic.severity, expectedDiagnostic.severity)
+    assert.strictEqual(actualDiagnostic.message, expectedDiagnostic.message)
+    assert.deepStrictEqual(actualDiagnostic.range, expectedDiagnostic.range)
+    assert.strictEqual(actualDiagnostic.severity, expectedDiagnostic.severity)
   })
 }
 
 suite('Should get linting', () => {
   const docUri = getDocUri('linting/missingArgument.prisma')
   const docUri2 = getDocUri('linting/wrongType.prisma')
-  const docUri3 = getDocUri('linting/requiredField.prisma')
 
   test('Diagnoses missing argument', async () => {
     await testDiagnostics(docUri, [
@@ -40,18 +39,7 @@ suite('Should get linting', () => {
       {
         message:
           'Type "Use" is neither a built-in type, nor refers to another model, custom type, or enum.',
-        range: toRange(14, 12, 14, 16),
-        severity: vscode.DiagnosticSeverity.Error,
-        source: '',
-      },
-    ])
-  })
-  test('Diagnoses required field', async () => {
-    await testDiagnostics(docUri3, [
-      {
-        message:
-          'Error validating: The relation field `author` uses the scalar fields authorId. At least one of those fields is required. Hence the relation field must be required as well.',
-        range: toRange(14, 2, 15, 0),
+        range: toRange(14, 12, 14, 15),
         severity: vscode.DiagnosticSeverity.Error,
         source: '',
       },
