@@ -6,14 +6,14 @@ function checkForUpdate({ channel }) {
     const currentPrismaVersion = readVersionFile({
       fileName: `prisma_${channel}`,
     })
-    console.log('current: ' + currentPrismaVersion)
+    console.log(`current '${channel}': ${currentPrismaVersion}`)
     const { stdout } = await execa('npm', [
       'show',
       `prisma@${channel}`,
       'version',
     ])
     const npmPrismaVersion = stdout
-    console.log('npm: ' + npmPrismaVersion)
+    console.log(`npm '${channel}': ${npmPrismaVersion}`)
     if (npmPrismaVersion === undefined) {
       throw Error('Could not get current Prisma CLI version.')
     }
