@@ -2,12 +2,14 @@ const fs = require('fs')
 const path = require('path')
 
 function patchBranchName() {
-  const prismaStableVersion = fs.readFileSync(path.join(__dirname, 'versions', './prisma_latest')).toString()
+  const prismaStableVersion = fs
+    .readFileSync(path.join(__dirname, 'versions', './prisma_latest'))
+    .toString()
   const tokens = prismaStableVersion.split('.')
   if (tokens.length !== 3) {
-      throw new Error(
-          `Version ${version} must have 3 tokens separated by "." character.`
-      )
+    throw new Error(
+      `Version ${version} must have 3 tokens separated by "." character.`,
+    )
   }
   // remove last digit
   const numbers = tokens.slice(0, 2)
@@ -16,19 +18,18 @@ function patchBranchName() {
   return patchName
 }
 
-function getBranchName({
-  branch_channel = ''
-}) {
+function getBranchName({ branch_channel = '' }) {
   switch (branch_channel) {
     case 'latest':
       return 'stable'
     case 'patch-dev':
       return patchBranchName()
     default:
-      throw new Error("Switching to another branch is only possible if on latest or patch-dev channel.")
+      throw new Error(
+        'Switching to another branch is only possible if on latest or patch-dev channel.',
+      )
   }
 }
-
 
 module.exports = { getBranchName }
 
